Modifying paragraph through buttons
Woithin this exercise I had to use a function to take acurrent font size, and add the value I get as an argument for the button that I am typing. I acheaved that creating four different .addEventListener that calls the function with the parameters I express in the function and I gave to the function as argument, the button incre +5 incre, etc. Now I have the problem that the function is already working buy it reset the value of currentFontSize, using the value that it has set by default that means the paragraph is unable to increment 1 by 1 from the real current size, it always does from the 12px value set as default...